VCBA-47 #time 4h adding support for more Verisense config options

+++ ShimmerDriver/src/main/java/com/shimmerresearch/bluetooth/SystemTimestampPlot.java
@@ -15,7 +15,7 @@ public class SystemTimestampPlot {
 	public ObjectCluster processSystemTimestampPlot(ObjectCluster objectCluster) {
 		if(!mIsFirstSystemTimestampOffsetStored) {
 			mIsFirstSystemTimestampOffsetStored = true;
-			long systemTimeStamp = objectCluster.mSystemTimeStamp;
+			double systemTimeStamp = objectCluster.mSystemTimeStamp;
 			mOffsetFirstTime = systemTimeStamp-objectCluster.getTimestampMilliSecs();
 		}
